New user-friendly functions! notes-to-collection! and notes-to-package!
The ability to be used with no setup? Check!
Lots of documentation with examples? Check!
The ability to be called with simple arguments? Check!
The ability to mix and match how complex you want to get? Check!

At a basic level, you can do this:

  (notes-to-package! ["Is this easy?" "Yep!"] "easy.apkg")

You can also dial up the complexity as you need more and more
functionality. The functions also work as a drop-in replacements for
`map-seq-to-package!` and `map-seq-to-collection!` (Though those are
still called internally, so I guess that's not too impressive)

But either way, spec conform is nice!